+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 3.1.2
+- Fix Misc and Mob Spawn Configs not generating sometimes
 
 # 3.1.1
 - Adjust BWG Blocks to match the same order in the creative tab as vanilla